Fix warning

+++ tests/integration/test_opendaq_device_modules/test_native_streaming_modules.cpp
@@ -1006,7 +1006,7 @@ TEST_F(NativeStreamingModulesTest, ParallelRpcCalls)
 
 TEST_F(NativeStreamingModulesTest, ParallelRpcCallsDefault)
 {
-    std::vector<size_t> propertyWriteHistory;
+    std::vector<Int> propertyWriteHistory;
 
     auto createServerInstance = [&propertyWriteHistory]()
     {
